+++ src/parser/enrichers/item/TomeOfLeadershipAndInfluence.mjs
@@ -1,0 +1,13 @@
+/* eslint-disable class-methods-use-this */
+import TomeOf from "./TomeOf.mjs";
+
+export default class TomeOfLeadershipAndInfluence extends TomeOf {
+
+  get effectData() {
+    return {
+      name: "Tome of Leadership and Influence",
+      ability: "cha",
+    };
+  }
+
+}